Finished code

+++ Lab 4/student-src/stats.c
@@ -17,8 +17,8 @@ stats_t stats;
     -----------------------------------------------------------------------------------
 */
 void compute_stats() {
-    stats.aat  = (stats.accesses * MEMORY_READ_TIME
+    stats.aat  = ((double) (stats.accesses * MEMORY_READ_TIME
     + stats.page_faults * DISK_PAGE_READ_TIME 
     + stats.writebacks * DISK_PAGE_WRITE_TIME) 
-    / stats.accesses;
+    / (double) stats.accesses);
 }
